client: consolidate node options with their methods
Move them together, as we did for most of the other options.
