+++ client/node_update.go
@@ -3,7 +3,15 @@ package client
 import (
 	"context"
 	"net/url"
+
+	"github.com/moby/moby/api/types/swarm"
 )
+
+// NodeUpdateOptions holds parameters to update nodes with.
+type NodeUpdateOptions struct {
+	Version swarm.Version
+	Node    swarm.NodeSpec
+}
 
 type NodeUpdateResult struct{}
