--- client/swarm_node_inspect_opts.go
@@ -1,4 +1,0 @@
-package client
-
-// NodeInspectOptions holds parameters to inspect nodes with.
-type NodeInspectOptions struct{}
