--- client/swarm_node_list_opts.go
@@ -1,6 +1,0 @@
-package client
-
-// NodeListOptions holds parameters to list nodes with.
-type NodeListOptions struct {
-	Filters Filters
-}
