--- client/swarm_node_remove_opts.go
@@ -1,6 +1,0 @@
-package client
-
-// NodeRemoveOptions holds parameters to remove nodes with.
-type NodeRemoveOptions struct {
-	Force bool
-}
