--- client/swarm_node_update_opts.go
@@ -1,9 +1,0 @@
-package client
-
-import "github.com/moby/moby/api/types/swarm"
-
-// NodeUpdateOptions holds parameters to update nodes with.
-type NodeUpdateOptions struct {
-	Version swarm.Version
-	Node    swarm.NodeSpec
-}
